image-generator/src/requirements.txt
numpy==2.4.0
Pillow==12.1.0
pillow_heif==1.1.1
